+++ abserv/abserv/Party.cpp
@@ -196,14 +196,14 @@ bool Party::IsLeader(Player* player) const
     return false;
 }
 
-uint8_t Party::GetPosition(Actor* actor)
+size_t Party::GetPosition(Actor* actor)
 {
     for (size_t i = 0; i < members_.size(); ++i)
     {
         if (auto sm = members_[i].lock())
         {
             if (sm->id_ == actor->id_)
-                return (static_cast<uint8_t>(i + 1));
+                return (i + 1);
         }
     }
     return 0;
