+++ abserv/abserv/Party.h
@@ -74,7 +74,7 @@ public:
         return members_[0].lock();
     }
     /// Get position of actor in party, 1-based, 0 = not found
-    uint8_t GetPosition(Actor* actor);
+    size_t GetPosition(Actor* actor);
     /// Tells all members to change the instance. The client will disconnect and reconnect to enter the instance.
     void ChangeInstance(const std::string& mapUuid);
